+++ src/client.spec.ts
@@ -1457,7 +1457,12 @@ describe('client', () => {
         .spyOn(client, 'fetch')
         .mockResolvedValue(mockResponse as Response);
       await expect(
-        client.getMap('vcs', '1.0.0', 'user-repos', 'github')
+        client.getMap({
+          name: 'user-repos',
+          provider: 'github',
+          scope: 'vcs',
+          version: '1.0.0',
+        })
       ).resolves.toEqual(mockResult);
       expect(fetchMock).toBeCalledTimes(1);
       expect(fetchMock).toBeCalledWith('/vcs/user-repos.github@1.0.0', {
@@ -1484,7 +1489,12 @@ describe('client', () => {
         .spyOn(client, 'fetch')
         .mockResolvedValue(mockResponse as Response);
       await expect(
-        client.getMap('vcs', '1.0.0', 'user-repos', 'github')
+        client.getMap({
+          name: 'user-repos',
+          provider: 'github',
+          scope: 'vcs',
+          version: '1.0.0',
+        })
       ).rejects.toEqual(new ServiceApiError(payload));
       expect(fetchMock).toBeCalledTimes(1);
       expect(fetchMock).toBeCalledWith('/vcs/user-repos.github@1.0.0', {
@@ -1583,7 +1593,12 @@ describe('client', () => {
         .spyOn(client, 'fetch')
         .mockResolvedValue(mockResponse as Response);
       await expect(
-        client.getMapSource('vcs', '1.0.0', 'user-repos', 'github')
+        client.getMapSource({
+          name: 'user-repos',
+          provider: 'github',
+          version: '1.0.0',
+          scope: 'vcs',
+        })
       ).resolves.toEqual('mapSource');
       expect(fetchMock).toBeCalledTimes(1);
       expect(fetchMock).toBeCalledWith('/vcs/user-repos.github@1.0.0', {
@@ -1610,7 +1625,12 @@ describe('client', () => {
         .spyOn(client, 'fetch')
         .mockResolvedValue(mockResponse as Response);
       await expect(
-        client.getMapSource('vcs', '1.0.0', 'user-repos', 'github')
+        client.getMapSource({
+          name: 'user-repos',
+          provider: 'github',
+          version: '1.0.0',
+          scope: 'vcs',
+        })
       ).rejects.toEqual(new ServiceApiError(payload));
       expect(fetchMock).toBeCalledTimes(1);
       expect(fetchMock).toBeCalledWith('/vcs/user-repos.github@1.0.0', {
@@ -1633,7 +1653,12 @@ describe('client', () => {
         .spyOn(client, 'fetch')
         .mockResolvedValue(mockResponse as Response);
       await expect(
-        client.getMapAST('vcs', '1.0.0', 'user-repos', 'github')
+        client.getMapAST({
+          name: 'user-repos',
+          provider: 'github',
+          version: '1.0.0',
+          scope: 'vcs',
+        })
       ).resolves.toEqual('mapAST');
       expect(fetchMock).toBeCalledTimes(1);
       expect(fetchMock).toBeCalledWith('/vcs/user-repos.github@1.0.0', {
@@ -1660,7 +1685,12 @@ describe('client', () => {
         .spyOn(client, 'fetch')
         .mockResolvedValue(mockResponse as Response);
       await expect(
-        client.getMapAST('vcs', '1.0.0', 'user-repos', 'github')
+        client.getMapAST({
+          name: 'user-repos',
+          provider: 'github',
+          scope: 'vcs',
+          version: '1.0.0',
+        })
       ).rejects.toEqual(new ServiceApiError(payload));
       expect(fetchMock).toBeCalledTimes(1);
       expect(fetchMock).toBeCalledWith('/vcs/user-repos.github@1.0.0', {